Renamed RelationDefinition files
